safari-600.7-branch/LayoutTests/svg/animations/svglength-animation-LengthModeOther-expected.txt

SVG 1.1 dynamic animation tests

Test SVGLength animation on LengthModeOther.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS circle.r.animVal.value is 10
PASS circle.r.baseVal.value is 10
PASS circle.r.animVal.value is 132.5
PASS circle.r.baseVal.value is 10
PASS circle.r.animVal.value is 254.9
PASS circle.r.baseVal.value is 10
PASS circle.r.animVal.value is 10
PASS circle.r.baseVal.value is 10
PASS successfullyParsed is true

TEST COMPLETE
